Make the popup open in the same window
