+++ index.html
@@ -11,7 +11,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Playfair+Display&family=Poppins&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="styles.css">
 </head>
-<body>
+<body id="body">
     <main id="main">
         <header class="header">
             <div class="row visible" id="nav-row">
@@ -64,6 +64,10 @@
         </section>
 
         <section class="work" id="portfolio">
+            <div class="big-container hidden" id="popup-window">
+                <div class="container" id="container">
+                </div>
+            </div>
         </section>
 
         <section class="about" id="about">
