--- proj-style.css
@@ -1,298 +1,0 @@
-*,
-::after,
-::before {
-  box-sizing: border-box;
-  padding: 0;
-  margin: 0;
-}
-
-a {
-  text-decoration: none;
-}
-
-ul {
-  list-style: none;
-}
-
-img {
-  max-width: 100%;
-}
-
-button {
-  cursor: pointer;
-  background: none;
-}
-
-.invisible {
-  display: none;
-}
-
-.container {
-  width: 90%;
-  margin: 5% auto;
-  max-width: 500px;
-  min-width: 240px;
-}
-
-.header {
-  padding-top: 2.5em;
-}
-
-.headerBtn {
-  float: right;
-  border: none;
-  width: 1.25em;
-  height: 1.25em;
-}
-
-.headerBtn img {
-  width: 100%;
-}
-
-.container .title {
-  font-family: 'Poppins', sans-serif;
-  font-style: normal;
-  font-weight: 700;
-  font-size: 1.25rem;
-  line-height: 80%;
-  color: #1a2236;
-}
-
-.workDesc-list {
-  display: flex;
-  gap: 10px;
-  margin: 23px 0 16px 0;
-}
-
-.workDesc-item {
-  font-family: 'Poppins', sans-serif;
-  font-style: normal;
-  font-weight: 300;
-  font-size: 0.646rem;
-  line-height: 80%;
-  color: #1a2236;
-  border: 0.516669px solid #1a2236;
-  padding: 0.375em 0.625em;
-}
-
-.grid {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  grid-template-rows: 80% 20%;
-  height: 202.7px;
-  width: 100%;
-  min-width: 278.75px;
-  max-width: 360px;
-}
-
-.grid img {
-  width: 100%;
-}
-
-.fig1 {
-  grid-column: 1 / span 4;
-  grid-row: 1 / 2;
-}
-
-.fig2 {
-  grid-column: 1 / 2;
-  grid-row: 2 / 3;
-}
-
-.fig3 {
-  grid-column: 2 / 3;
-  grid-row: 2 / 3;
-}
-
-.fig4 {
-  grid-column: 3 / 4;
-  grid-row: 2 / 3;
-}
-
-.fig5 {
-  grid-column: 4 / 5;
-  grid-row: 2 / 3;
-}
-
-.detail-desc p {
-  font-family: 'Poppins', sans-serif;
-  font-style: normal;
-  font-weight: 400;
-  font-size: 1.125rem;
-  line-height: 27px;
-  color: #1a2236;
-  margin: 29px 0;
-}
-
-.btn-cont {
-  display: flex;
-  flex-direction: column;
-  gap: 29px;
-  width: 80%;
-  margin: 0 auto;
-  max-width: 350px;
-}
-
-.btn-cont button {
-  font-family: 'Poppins', sans-serif;
-  font-style: normal;
-  font-weight: 500;
-  font-size: 1.125rem;
-  line-height: 27px;
-  text-align: center;
-  color: #1a2236;
-  background: #fff;
-  border: 1px solid #1a2236;
-  padding: 0.75em 0;
-  background-position: 90% 50%;
-  background-repeat: no-repeat;
-  cursor: pointer;
-}
-
-.btn-cont button:hover {
-  background: #1a2236;
-  color: #fff;
-  border: 1px solid #fff;
-}
-
-.btn-cont button:active {
-  border: 2px solid #fff;
-}
-
-.btn-cont button:first-of-type {
-  background-image: url(img/Union2.svg);
-}
-
-.btn-cont button:last-of-type {
-  background-image: url(img/Vector.svg);
-}
-
-@media screen and (min-width: 768px) {
-  .invisible {
-    display: block;
-  }
-
-  .visible {
-    display: none;
-  }
-
-  .big-container {
-    background: rgba(26, 34, 54, 0.95);
-    backdrop-filter: blur(12px);
-    max-width: 1640px;
-    padding: 4% 0;
-    margin: 0 auto;
-  }
-
-  .container {
-    width: 90%;
-    margin: 0 auto;
-    background: #fff;
-    box-shadow: 0 10px 20px #00000040;
-    height: 1720px;
-    max-width: 1440px;
-  }
-
-  .small-container {
-    width: 90%;
-    margin: 0 auto;
-    padding: 5% 0;
-  }
-
-  .header {
-    padding: 0;
-  }
-
-  .headerBtn {
-    width: 32px;
-    height: 32px;
-  }
-
-  .headerBtn img {
-    width: 100%;
-  }
-
-  .container .title {
-    font-family: 'Poppins', sans-serif;
-    font-style: normal;
-    font-weight: 700;
-    font-size: 3.75rem;
-    line-height: 80%;
-  }
-
-  .workDesc-list {
-    margin: 32px 0 40px 0;
-    gap: 20px;
-  }
-
-  .workDesc-item {
-    padding: 0.75em 1.25em;
-    font-weight: 300;
-    font-size: 1.25rem;
-    line-height: 80%;
-  }
-
-  .grid {
-    height: 657.32px;
-    width: 100%;
-    max-width: 1340px;
-    margin: 0 auto;
-  }
-
-  .detail-desc {
-    width: 60%;
-    margin: 0 auto;
-    margin-top: 68.68px;
-  }
-
-  .detail-desc p {
-    font-weight: 400;
-    font-size: 1.125rem;
-    line-height: 27px;
-    margin-bottom: 40px;
-  }
-
-  .btn-cont {
-    width: 80%;
-    flex-direction: row;
-    justify-content: flex-start;
-    margin: 0;
-    max-width: 750px;
-  }
-
-  .btn-cont button {
-    width: 100%;
-    text-align: left;
-    padding-left: 1.5625em;
-  }
-
-  .direction-btn {
-    display: flex;
-    justify-content: space-between;
-    margin-top: 120px;
-  }
-
-  .direction-btn button {
-    font-family: 'Poppins', sans-serif;
-    font-style: normal;
-    font-weight: 300;
-    font-size: 1.125rem;
-    line-height: 27px;
-    color: #1a2236;
-    background-color: #fff;
-    border: none;
-    background-repeat: no-repeat;
-  }
-
-  .direction-btn button:first-of-type {
-    padding-left: 2.5em;
-    background-image: url(img/Union3.svg);
-    background-position: left center;
-  }
-
-  .direction-btn button:last-of-type {
-    padding-right: 2.5em;
-    background-image: url(img/Union4.svg);
-    background-position: right center;
-  }
-}
