--- proj4-desc.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="shortcut icon" type="image/svg" href="img/image_geometry_1.svg">
-    <title>Project 4</title>
-    <link rel="stylesheet" href="proj-style.css">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Poppins&display=swap" rel="stylesheet">
-</head>
-<body>
-    <div class="big-container">
-        <div class="container" id="container">
-        </div>
-    </div>
-
-    <script type="module" src="script-desc4.js"></script>
-</body>
-</html>
