+++ styles.css
@@ -252,6 +252,189 @@ main {
   align-items: center;
 }
 
+/* project description popup style */
+.blur {
+  overflow-y: hidden;
+}
+
+#popup-window {
+  z-index: 15;
+  height: 100%;
+  width: 100%;
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  background: rgba(193, 199, 208, 0.7);
+  backdrop-filter: blur(4px);
+  overflow-y: auto;
+  margin: auto 0;
+}
+
+.big-container a {
+  text-decoration: none;
+}
+
+.big-container ul {
+  list-style: none;
+}
+
+.big-container img {
+  max-width: 100%;
+}
+
+.big-container button {
+  cursor: pointer;
+  background: none;
+}
+
+.big-container .invisible {
+  display: none;
+}
+
+.big-container .container {
+  width: 90%;
+  margin: 5% auto;
+  max-width: 500px;
+  min-width: 240px;
+}
+
+.big-container .header {
+  padding-top: 2.5em;
+}
+
+.big-container .headerBtn {
+  float: right;
+  border: none;
+  width: 1.25em;
+  height: 1.25em;
+}
+
+.big-container .headerBtn img {
+  width: 100%;
+}
+
+.big-container .container .title {
+  font-family: 'Poppins', sans-serif;
+  font-style: normal;
+  font-weight: 700;
+  font-size: 1.25rem;
+  line-height: 80%;
+  color: #1a2236;
+}
+
+.big-container .workDesc-list {
+  display: flex;
+  gap: 10px;
+  margin: 23px 0 16px 0;
+}
+
+.big-container .workDesc-item {
+  font-family: 'Poppins', sans-serif;
+  font-style: normal;
+  font-weight: 300;
+  font-size: 0.646rem;
+  line-height: 80%;
+  color: #1a2236;
+  border: 0.516669px solid #1a2236;
+  padding: 0.375em 0.625em;
+}
+
+.big-container .grid {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  grid-template-rows: 80% 20%;
+  height: 202.7px;
+  width: 100%;
+  min-width: 278.75px;
+  max-width: 360px;
+}
+
+.big-container .grid img {
+  width: 100%;
+}
+
+.big-container .fig1 {
+  grid-column: 1 / span 4;
+  grid-row: 1 / 2;
+}
+
+.big-container .fig2 {
+  grid-column: 1 / 2;
+  grid-row: 2 / 3;
+}
+
+.big-container .fig3 {
+  grid-column: 2 / 3;
+  grid-row: 2 / 3;
+}
+
+.big-container .fig4 {
+  grid-column: 3 / 4;
+  grid-row: 2 / 3;
+}
+
+.big-container .fig5 {
+  grid-column: 4 / 5;
+  grid-row: 2 / 3;
+}
+
+.big-container .detail-desc p {
+  font-family: 'Poppins', sans-serif;
+  font-style: normal;
+  font-weight: 400;
+  font-size: 1.125rem;
+  line-height: 27px;
+  color: #1a2236;
+  margin: 29px 0;
+}
+
+.big-container .btn-cont {
+  display: flex;
+  flex-direction: column;
+  gap: 29px;
+  width: 80%;
+  margin: 0 auto;
+  max-width: 350px;
+}
+
+.big-container .btn-cont button {
+  font-family: 'Poppins', sans-serif;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 1.125rem;
+  line-height: 27px;
+  text-align: center;
+  color: #1a2236;
+  background: #fff;
+  border: 1px solid #1a2236;
+  padding: 0.75em 0;
+  background-position: 90% 50%;
+  background-repeat: no-repeat;
+  cursor: pointer;
+}
+
+.big-container .btn-cont button:hover {
+  background: #1a2236;
+  color: #fff;
+  border: 1px solid #fff;
+}
+
+.big-container .btn-cont button:active {
+  border: 2px solid #fff;
+}
+
+.big-container .btn-cont button:first-of-type {
+  background-image: url(img/Union2.svg);
+}
+
+.big-container .btn-cont button:last-of-type {
+  background-image: url(img/Vector.svg);
+}
+
+/* project description style end */
+
 .skill-pack {
   display: flex;
   flex-direction: column;
@@ -726,6 +909,138 @@ label:hover {
   .workBtn:active {
     border: 2.5px solid #fff;
   }
+
+  /* project description popup style */
+
+  .big-container .invisible {
+    display: block;
+  }
+
+  .big-container .visible {
+    display: none;
+  }
+
+  .big-container {
+    background: rgba(26, 34, 54, 0.95);
+    backdrop-filter: blur(12px);
+    max-width: 1640px;
+    padding: 4% 0;
+    margin: 0 auto;
+  }
+
+  .big-container .container {
+    width: 90%;
+    margin: 0 auto;
+    background: #fff;
+    box-shadow: 0 10px 20px #00000040;
+    height: 1720px;
+    max-width: 1440px;
+  }
+
+  .big-container .small-container {
+    width: 90%;
+    margin: 0 auto;
+    padding: 5% 0;
+  }
+
+  .big-container .header {
+    padding: 0;
+  }
+
+  .big-container .headerBtn {
+    width: 32px;
+    height: 32px;
+  }
+
+  .big-container .headerBtn img {
+    width: 100%;
+  }
+
+  .big-container .container .title {
+    font-family: 'Poppins', sans-serif;
+    font-style: normal;
+    font-weight: 700;
+    font-size: 3.75rem;
+    line-height: 80%;
+  }
+
+  .big-container .workDesc-list {
+    margin: 32px 0 40px 0;
+    gap: 20px;
+  }
+
+  .big-container .workDesc-item {
+    padding: 0.75em 1.25em;
+    font-weight: 300;
+    font-size: 1.25rem;
+    line-height: 80%;
+  }
+
+  .big-container .grid {
+    height: 757.32px;
+    width: 100%;
+    max-width: 1340px;
+    margin: 0 auto;
+  }
+
+  .big-container .detail-desc {
+    width: 60%;
+    margin: 0 auto;
+    margin-top: 68.68px;
+  }
+
+  .big-container .detail-desc p {
+    font-weight: 400;
+    font-size: 1.125rem;
+    line-height: 27px;
+    margin-bottom: 40px;
+  }
+
+  .big-container .btn-cont {
+    width: 80%;
+    flex-direction: row;
+    justify-content: flex-start;
+    margin: 0;
+    max-width: 750px;
+  }
+
+  .big-container .btn-cont button {
+    width: 100%;
+    text-align: left;
+    padding-left: 1.5625em;
+  }
+
+  .big-container .direction-btn {
+    display: flex;
+    justify-content: space-between;
+    margin-top: 120px;
+  }
+
+  .big-container .direction-btn button {
+    font-family: 'Poppins', sans-serif;
+    font-style: normal;
+    font-weight: 300;
+    font-size: 1.125rem;
+    line-height: 27px;
+    color: #1a2236;
+    background-color: #fff;
+    border: none;
+    background-repeat: no-repeat;
+  }
+
+  .big-container .direction-btn button:first-of-type {
+    padding-left: 2.5em;
+    background-image: url(img/Union3.svg);
+    background-position: left center;
+  }
+
+  .big-container .direction-btn button:last-of-type {
+    padding-right: 2.5em;
+    background-image: url(img/Union4.svg);
+    background-position: right center;
+  }
+
+  /* project description popup style end */
 
   .about {
     display: flex;
